Refactor: Remove authentication, add hero note, and implement popup warning
