--- trackFlow/src/authentication/authMethods.jsx
@@ -1,25 +1,0 @@
-import { auth } from './firebaseConfig';
-import { GoogleAuthProvider, createUserWithEmailAndPassword, signInWithEmailAndPassword, signInWithPopup } from 'firebase/auth';
-
-// Creating User Profile
-export const doCreateUserWithEmailAndPassword = async (email, password) => {
-    return createUserWithEmailAndPassword(auth, email, password);
-};
-
-// Sign in with Email 
-export const doSignInWithEmailAndPassword = async (email, password) => {
-    return signInWithEmailAndPassword(auth, email, password);
-};
-
-// Sign in with Google
-export const doSignInWithGoogle = async () => {
-    const provider = new GoogleAuthProvider();
-    const result = await signInWithPopup(auth, provider);
-    
-    return result;
-};
-
-// Sign Out
-export const doSignOut = () => {
-    return auth.signOut();
-};
