--- trackFlow/src/context/userContext.jsx
@@ -1,33 +1,0 @@
-// src/context/UserContext.jsx
-import React, { createContext, useState, useEffect } from 'react';
-
-const UserContext = createContext();
-
-const UserProvider = ({ children }) => {
-  const [user, setUser] = useState(null);
-
-  useEffect(() => {
-    const storedUser = JSON.parse(localStorage.getItem('user'));
-    if (storedUser) {
-      setUser(storedUser);
-    }
-  }, []);
-
-  const loginUser = (userData) => {
-    setUser(userData);
-    localStorage.setItem('user', JSON.stringify(userData));
-  };
-
-  const logoutUser = () => {
-    setUser(null);
-    localStorage.removeItem('user');
-  };
-
-  return (
-    <UserContext.Provider value={{ user, loginUser, logoutUser }}>
-      {children}
-    </UserContext.Provider>
-  );
-};
-
-export { UserContext, UserProvider };
