+++ trackFlow/src/main.jsx
@@ -1,13 +1,10 @@
 import { StrictMode } from 'react';
 import { createRoot } from 'react-dom/client';
-import { UserProvider } from './context/userContext.jsx';
 import App from './App.jsx'
 import './index.css'
 
 createRoot(document.getElementById('root')).render(
   <StrictMode>
-    <UserProvider>
-      <App />
-    </UserProvider>
+    <App />
   </StrictMode>,
 )
